refactor: follow single responsibility principle
